Complete api and data imports'

+++ db/migrate/20190207221747_create_players.rb
@@ -6,9 +6,9 @@ class CreatePlayers < ActiveRecord::Migration[5.2]
       t.integer :games_played
       t.integer :game_minutes
       t.integer :points
-      t.integer :FG_percentage
-      t.integer :three_point_percentage
-      t.integer :FT_percentage
+      t.float :FG_percentage
+      t.float :three_point_percentage
+      t.float :FT_percentage
       t.timestamps
     end
   end
